+++ lib/data/mocks/image_mock_data.dart
@@ -1,0 +1,5 @@
+// data/mocks/image_mock_data.dart
+import '../../domain/models/image_model.dart';
+import '../models/image_model.dart';
+
+final List<ImageModel> mockImageList = imgList;
