+++ lib/data/models/image_model.dart
@@ -1,0 +1,64 @@
+import '../../domain/models/image_model.dart';
+
+final List<ImageModel> imgList = [
+  ImageModel(
+      dateOfCrete: '01.03.2023',
+      imagePath: 'assets/images/Default_madoka_magica_art_0.jpg'),
+  ImageModel(
+      dateOfCrete: '02.05.2023',
+      imagePath: 'assets/images/Default_madoka_magica_art_1.jpg'),
+  ImageModel(
+      dateOfCrete: '03.06.2023',
+      imagePath: 'assets/images/Default_madoka_magica_art_2.jpg'),
+  ImageModel(
+      dateOfCrete: '05.06.2023',
+      imagePath: 'assets/images/Default_madoka_magica_art_3.jpg'),
+  ImageModel(
+      dateOfCrete: '09.06.2023',
+      imagePath:
+          'assets/images/Default_madoka_magica_from_anime_Madoka_Magica_0.jpg'),
+  ImageModel(
+      dateOfCrete: '02.07.2023',
+      imagePath:
+          'assets/images/Default_madoka_magica_from_anime_Madoka_Magica_1.jpg'),
+  ImageModel(
+      dateOfCrete: '03.07.2023',
+      imagePath:
+          'assets/images/Default_madoka_magica_from_anime_Madoka_Magica_2.jpg'),
+  ImageModel(
+      dateOfCrete: '01.08.2023',
+      imagePath:
+          'assets/images/Default_madoka_magica_from_anime_Madoka_Magica_3.jpg'),
+  ImageModel(
+      dateOfCrete: '01.08.2023',
+      imagePath:
+          'assets/images/Default_madoka_magica_from_anime_Madoka_Magica_with_bow_0.jpg'),
+  ImageModel(
+      dateOfCrete: '01.03.2023',
+      imagePath:
+          'assets/images/Default_madoka_magica_from_anime_Madoka_Magica_with_bow_1.jpg'),
+  ImageModel(
+      dateOfCrete: '01.03.2023',
+      imagePath:
+          'assets/images/Default_madoka_magica_from_anime_Madoka_Magica_with_bow_2.jpg'),
+  ImageModel(
+      dateOfCrete: '01.03.2023',
+      imagePath:
+          'assets/images/Default_madoka_magica_from_anime_Madoka_Magica_with_bow_3.jpg'),
+  ImageModel(
+      dateOfCrete: '01.03.2023',
+      imagePath:
+          'assets/images/Default_madoka_magica_from_anime_Madoka_Magica_with_bow_in_spa_0.jpg'),
+  ImageModel(
+      dateOfCrete: '01.03.2023',
+      imagePath:
+          'assets/images/Default_madoka_magica_from_anime_Madoka_Magica_with_bow_in_spa_1.jpg'),
+  ImageModel(
+      dateOfCrete: '01.03.2023',
+      imagePath:
+          'assets/images/Default_madoka_magica_from_anime_Madoka_Magica_with_bow_in_spa_2.jpg'),
+  ImageModel(
+      dateOfCrete: '01.03.2023',
+      imagePath:
+          'assets/images/Default_madoka_magica_from_anime_Madoka_Magica_with_bow_in_spa_3.jpg'),
+];
