+++ lib/domain/models/image_model.dart
@@ -1,0 +1,6 @@
+class ImageModel {
+  final String dateOfCrete;
+  final String imagePath;
+
+  ImageModel({required this.dateOfCrete, required this.imagePath});
+}
