+++ lib/uikit/theme/theme_data.dart
@@ -13,8 +13,6 @@ abstract class AppThemeData {
       onSecondary: _lightColorScheme.onSecondary,
       error: _lightColorScheme.danger,
       onError: _lightColorScheme.onDanger,
-      background: _lightColorScheme.background,
-      onBackground: _lightColorScheme.onBackground,
       surface: _lightColorScheme.surface,
       onSurface: _lightColorScheme.onSurface,
     ),
@@ -49,8 +47,6 @@ abstract class AppThemeData {
       onSecondary: _darkColorScheme.onSecondary,
       error: _darkColorScheme.danger,
       onError: _darkColorScheme.onDanger,
-      background: _darkColorScheme.background,
-      onBackground: _darkColorScheme.onBackground,
       surface: _darkColorScheme.surface,
       onSurface: _darkColorScheme.onSurface,
     ),
